[SPC] Add payeeName option
Implementation for the recent spec change[1] to add a payeeName option
as an optional alternative to payeeOrigin. The API now accepts one or
both of payeeName and payeeOrigin. If both are provided, they are both
shown in the UI with the format "payeeName (payeeOrigin)".

This formatting when both fields are present is done in the UI code for
two reasons; it's easier to test, and it can be considered a UI
implementation detail how they are displayed, e.g. on a different
platform or future iteration of the UI the origin could be displayed
in a different font than the name, so the MVC model keeps them separate.

[1] https://github.com/w3c/secure-payment-confirmation/pull/180

Bug: 1298505
Change-Id: I19a255ce9f162b7d9e6c4cf0c2164242fa96cfc3