[FlexNG] Handle break rules for multi-line columns
Add support for break-before/after/inside rules for multi-line column
flex containers. Because we fragment the columns independently, we
needed a way to keep track of the break-after and early-break for each
column (tracked by ColumnBreakInfo).

We also may now have more than one early break in a given fragment, so
a way to pass in additional early breaks was added to the layout
algorithm when we abort due to an early break.

Because columns can be considered a "row" of columns, the break-before
for all first items in a column were accumulated to the container
(same with break-after and the last items in each column).

Bug: 660611
Change-Id: I198fe0543f2c89bb5d188a4e2996bec89b4546cf